Slight Marker update
-seperated window logic from home.tsx into InfoWindow.tsx
-fixed window opening and closing animations
-limited the window height to sit bellow the notch
-added a blurred background when the window is fully maximised
